Improve and split out range limited stopwatch clock
